A class for peptide-spectrum matches results

The PSM class is a simple class to store and manipulate peptide-spectrum matches. The class encapsulates PSM data as a DataFrame (or more specifically a DFrame) with additional lightweight metadata annotation.

There are two types of PSM objects:

Objects with duplicated spectrum identifiers. This holds for multiple matches to the same spectrum, be it different peptide sequences or the same sequence with or without a post-translational modification. Such objects are typically created with the PSM() constructor starting from mzIdentML files.

Reduced object where the spectrum identifiers (or any equivalent column - see example below) are unique keys within the PSM table. Matches to the same scan/spectrum in each raw data file are merged into a single data row. Reduced PSM object are created with the reducePSMs() function. See examples below.

Objects can be checked for their reduced state with the reduced() function which returns TRUE for reduced instances, FALSE when the spectrum identifiers are duplicated, or NA when unknown. The flag can also be set explicitly with the reduced()<- setter.

The constructor uses parsers provided by the mzR or mzID packages to read the mzIdentML data. See the vignette for some apparent differences in their outputs.

PSM(files, parser = c("mzR", "mzID"), BPPARAM = SerialParam())

reduced(x, spectrumID)

reduced(x) <- value

reducePSMs(x, k)

Arguments

files
character() of mzid file names.

parser
character(1) defining the parser to be used to read the mzIdentML files. One of "mzR" (default) or "mzID".

BPPARAM
an object inheriting from BiocParallelParam to control parallel processing. The default value is SerialParam() to read files in series.

x
An instance of class PSM.

spectrumID
character(1) with the column name referring to the spectrum identifier or any other unique key for the PSM object x. This parameter can be set to calculate the reduced state explicitly by by-passing the reduced metadata flag.

value
logical(1) to set the reduced state of the PSM object.

k
A vector or factor of length equal to nrow(x) that defines the primary key used to reduce x. This typically corresponds to the spectrum identifier, often names spectrumID.

Value

PSM() returns a PSM object containing the peptide-spectrum matches stored in files.

reducePSMs() returns a reduced version of the x input.

Examples

## ---------------------------------
## Example with a single mzid file
## ---------------------------------

f <- msdata::ident(full.names = TRUE, pattern = "TMT")
basename(f)

#> [1] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzid"

## mzR parser (default)
psm <- PSM(f)

#> Loading required namespace: mzR

psm

#> PSM with 5802 rows and 35 columns.
#> names(35): sequence spectrumID ... subReplacementResidue subLocation

## mzID parser
PSM(f, parser = "mzID")

#> Loading required namespace: mzID

#> reading TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzid...

#> Warning: 'as.is' should be specified by the caller; using TRUE

#> Warning: 'as.is' should be specified by the caller; using TRUE

#> Warning: 'as.is' should be specified by the caller; using TRUE

#> Warning: 'as.is' should be specified by the caller; using TRUE

#> Warning: 'as.is' should be specified by the caller; using TRUE

#> Warning: 'as.is' should be specified by the caller; using TRUE

#> Warning: 'as.is' should be specified by the caller; using TRUE

#> Warning: 'as.is' should be specified by the caller; using TRUE

#> Warning: 'as.is' should be specified by the caller; using TRUE

#> Warning: 'as.is' should be specified by the caller; using TRUE

#> Warning: 'as.is' should be specified by the caller; using TRUE

#> Warning: 'as.is' should be specified by the caller; using TRUE

#> Warning: 'as.is' should be specified by the caller; using TRUE

#>  DONE!

#> PSM with 5759 rows and 30 columns.
#> names(30): spectrumid scan number(s) ... spectrumFile databaseFile

i <- which(duplicated(psm$spectrumID))[1:2]
i <- which(psm$spectrumID %in% psm$spectrumID[i])
psm2 <- psm[i, ]
reduced(psm2)

#> [1] NA

## Peptide sequence CIDRARHVEVQIFGDGKGRVVALGERDCSLQRR with
## Carbamidomethyl modifications at positions 1 and 28.
DataFrame(psm2[, c("sequence", "spectrumID", "modName", "modLocation")])

#> DataFrame with 6 rows and 4 columns
#>                 sequence             spectrumID         modName modLocation
#>              <character>            <character>     <character>   <integer>
#> 1 CIDRARHVEVQIFGDGKGRV.. controllerType=0 con.. Carbamidomethyl           1
#> 2 CIDRARHVEVQIFGDGKGRV.. controllerType=0 con.. Carbamidomethyl          28
#> 3       KCNQCLKVACTLFYCK controllerType=0 con.. Carbamidomethyl           2
#> 4       KCNQCLKVACTLFYCK controllerType=0 con.. Carbamidomethyl           5
#> 5       KCNQCLKVACTLFYCK controllerType=0 con.. Carbamidomethyl          10
#> 6       KCNQCLKVACTLFYCK controllerType=0 con.. Carbamidomethyl          15

reduced(psm2) <- FALSE

rpsm2 <- reducePSMs(psm2, psm2$spectrumID)
rpsm2

#> Reduced PSM with 2 rows and 35 columns.
#> names(35): sequence spectrumID ... subReplacementResidue subLocation

DataFrame(rpsm2[, c("sequence", "spectrumID", "modName", "modLocation")])

#> DataFrame with 2 rows and 4 columns
#>                                                    sequence    spectrumID
#>                                                 <character>   <character>
#> controllerType=0 controllerNumber=1 scan=5490 KCNQCLKVAC... controller...
#> controllerType=0 controllerNumber=1 scan=6518 CIDRARHVEV... controller...
#>                                                     modName   modLocation
#>                                                 <character> <IntegerList>
#> controllerType=0 controllerNumber=1 scan=5490 Carbamidom...    2,5,10,...
#> controllerType=0 controllerNumber=1 scan=6518 Carbamidom...          1,28

reduced(rpsm2)

#> [1] TRUE

## ---------------------------------
## Multiple mzid files
## ---------------------------------

library(rpx)
PXD022816 <- PXDataset("PXD022816")
PXD022816

#> Object of class "PXDataset"
#>  Id: PXD022816 with 32 files
#>  [1] 'QEP2LC6_HeLa_50ng_251120_01-calib.mzID.gz' ... [32] 'checksum.txt'
#>  Use 'pxfiles(.)' to see all files.

(mzids <- pxget(PXD022816, grep("mzID", pxfiles(PXD022816))[1:2]))

#> Loading QEP2LC6_HeLa_50ng_251120_01-calib.mzID.gz from cache.

#> Loading QEP2LC6_HeLa_50ng_251120_02-calib.mzID.gz from cache.

#> [1] "~/.cache/rpx/8dc56596c3f_QEP2LC6_HeLa_50ng_251120_01-calib.mzID.gz"
#> [2] "~/.cache/rpx/8dc3f02b93_QEP2LC6_HeLa_50ng_251120_02-calib.mzID.gz" 

psm <- PSM(mzids)
psm

#> PSM with 50314 rows and 31 columns.
#> names(31): sequence spectrumID ... subReplacementResidue subLocation

## Here, spectrum identifiers are repeated accross files
psm[grep("scan=20000", psm$spectrumID), "spectrumFile"]

#> [1] "P:\\Projects\\Morgenstern\\HF2_QC_for_the_paper\\QEP2\\2020-11-29-13-56-12\\Task1-CalibrateTask\\QEP2LC6_HeLa_50ng_251120_01-calib.mzML"
#> [2] "P:\\Projects\\Morgenstern\\HF2_QC_for_the_paper\\QEP2\\2020-11-29-13-56-12\\Task1-CalibrateTask\\QEP2LC6_HeLa_50ng_251120_02-calib.mzML"

## Let's create a new primary identifier composed of the scan
## number and the file name
psm$pkey <- paste(sub("^.+Task\\\\", "", psm$spectrumFile),
                  sub("^.+scan=", "", psm$spectrumID),
                  sep = "::")
head(psm$pkey)

#> [1] "QEP2LC6_HeLa_50ng_251120_01-calib.mzML::3426" 
#> [2] "QEP2LC6_HeLa_50ng_251120_01-calib.mzML::20165"
#> [3] "QEP2LC6_HeLa_50ng_251120_01-calib.mzML::20180"
#> [4] "QEP2LC6_HeLa_50ng_251120_01-calib.mzML::15180"
#> [5] "QEP2LC6_HeLa_50ng_251120_01-calib.mzML::10327"
#> [6] "QEP2LC6_HeLa_50ng_251120_01-calib.mzML::12894"

## the PSM is not reduced
reduced(psm, "pkey")

#> [1] FALSE

DataFrame(psm[6:7, ])

#> DataFrame with 2 rows and 32 columns
#>        sequence    spectrumID chargeState      rank passThreshold
#>     <character>   <character>   <integer> <integer>     <logical>
#> 1 SEGSSCALES... controller...           4         1          TRUE
#> 2 SEGSSCALES... controller...           4         1          TRUE
#>   experimentalMassToCharge calculatedMassToCharge  peptideRef    modNum
#>                  <numeric>              <numeric> <character> <integer>
#> 1                  796.367                796.367        P_19         2
#> 2                  796.367                796.367        P_19         2
#>     isDecoy        post         pre     start       end DatabaseAccess
#>   <logical> <character> <character> <integer> <integer>    <character>
#> 1     FALSE           I           K       220       251         Q58D62
#> 2     FALSE           I           K       220       251         Q58D62
#>   DBseqLength   DatabaseSeq DatabaseDescription scan.start.time acquisitionNum
#>     <integer>   <character>         <character>     <character>      <numeric>
#> 1         387 MNVLLLLVLC...       Q58D62|FET...       32.788864          12894
#> 2         387 MNVLLLLVLC...       Q58D62|FET...       32.788864          12894
#>     spectrumFile        idFile MetaMorpheus.score PSM.level.q.value
#>      <character>   <character>          <numeric>         <numeric>
#> 1 P:\\Project... 8dc56596c3...            31.2586                 0
#> 2 P:\\Project... 8dc56596c3...            31.2586                 0
#>   modPeptideRef       modName   modMass modLocation subOriginalResidue
#>     <character>   <character> <numeric>   <integer>        <character>
#> 1          P_19 Carbamidom...   57.0215           6                 NA
#> 2          P_19 Carbamidom...   57.0215          19                 NA
#>   subReplacementResidue subLocation          pkey
#>             <character>   <integer>   <character>
#> 1                    NA          NA QEP2LC6_He...
#> 2                    NA          NA QEP2LC6_He...

## same sequence, same spectrumID, same file
psm$sequence[6:7]

#> [1] "SEGSSCALESPGSVPVGICHGSLGEPQGNQGK" "SEGSSCALESPGSVPVGICHGSLGEPQGNQGK"

psm$pkey[6:7]

#> [1] "QEP2LC6_HeLa_50ng_251120_01-calib.mzML::12894"
#> [2] "QEP2LC6_HeLa_50ng_251120_01-calib.mzML::12894"

## different modification locations
psm$modLocation[6:7]

#> [1]  6 19

rpsm <- reducePSMs(psm, psm$pkey)
rpsm

#> Reduced PSM with 32739 rows and 32 columns.
#> names(32): sequence spectrumID ... subLocation pkey

reduced(rpsm, "pkey")

#> [1] TRUE

## the two rows are now merged into a single one; the distinct
## modification locations are preserved.
(i <- which(rpsm$pkey == "QEP2LC6_HeLa_50ng_251120_01-calib.mzML::12894"))

#> [1] 2343

DataFrame(rpsm[i, c("sequence", "pkey", "modName", "modLocation")])

#> DataFrame with 1 row and 4 columns
#>                                                      sequence          pkey
#>                                               <CharacterList>   <character>
#> QEP2LC6_HeLa_50ng_251120_01-calib.mzML::12894   SEGSSCALES... QEP2LC6_He...
#>                                                       modName   modLocation
#>                                               <CharacterList> <IntegerList>
#> QEP2LC6_HeLa_50ng_251120_01-calib.mzML::12894   Carbamidom...          6,19