Filter out unreliable PSMs.

Functions to filter out PSMs matching. The PSMs should be stored in a PSM such as those produced by PSM().

filterPsmDecoy() filters out decoy PSMs, i.e. those annotated as isDecoy.

filterPsmRank() filters out PSMs of rank > 1.

filterPsmNonProteotypic() filters out non-proteotypic PSMs, i.e. those that match multiple proteins.

filterPsmBestScore() filters out PSMs but those with he highest score when matching the same spectrum identifier.

filterPsmMods() filters out PSMs that contain any PTM, as defined by non-missing (i.e. non-NA) modification name.

filterPSMs(
  x,
  decoy = "isDecoy",
  rank = "rank",
  accession = "DatabaseAccess",
  spectrumID = "spectrumID",
  verbose = TRUE
)

filterPsmDecoy(x, decoy = "isDecoy", verbose = TRUE)

filterPsmRank(x, rank = "rank", verbose = TRUE)

filterPsmNonProteotypic(
  x,
  accession = "DatabaseAccess",
  spectrumID = "spectrumID",
  verbose = TRUE
)

filterPsmBestScore(x, score, spectrumID = "spectrumID", verbose = TRUE)

filterPsmMods(x, mod = "modName", verbose = TRUE)

Arguments

x
An instance of class PSM.

decoy
character(1) with the column name specifying whether entries match the decoy database or not. Default is "isDecoy". The column should be a logical and only PSMs holding a FALSE are retained. Filtering is ignored if set to NULL.

rank
character(1) with the column name holding the rank of the PSM. Default is "rank". This column should be a numeric and only PSMs having rank equal to 1 are retained. Filtering is ignored if set to NULL.

accession
character(1) with the column name holding the protein (groups) accession. Default is "DatabaseAccess". Filtering is ignored if set to NULL.

spectrumID
character(1) with the name of the spectrum identifier column. Default is spectrumID. Filtering is ignored if set to NULL.

verbose
logical(1) setting the verbosity flag.

score
character(1) with the name of a score column to be used. Filtering is ignored if set to NULL.

mod
character(1) with the column name holding the modification name. Default is "modName". Filtering is ignored if set to NULL.

Value

A new filtered PSM object with the same columns as the input x.

Author

Laurent Gatto

Examples

f <- msdata::ident(full.names = TRUE, pattern = "TMT")
basename(f)

#> [1] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzid"

id <- PSM(f)
filterPSMs(id)

#> Starting with 5802 PSMs:

#>  removed 2896 decoy hits

#>  removed 155 PSMs with rank > 1

#>  removed 85 non-proteotypic peptides

#> 2666 PSMs left.

#> PSM with 2666 rows and 35 columns.
#> names(35): sequence spectrumID ... subReplacementResidue subLocation