Import peptide-spectrum matches

Reads as set of mzId files containing PSMs and returns the PSMs as a DataFrame.

readPSMs(files, parser = c("mzR", "mzID"))

Arguments

files
A character of mzid files.

parser
character(1) defining the parser to be used to read the mzIdentML files. One of "mzR" (default) or "mzID".

Value

A DataFrame containing the PSMs stored in the mzId files.

Details

This function uses parsers provided by the mzR or mzID packages to read the mzIdentML data. See the vignette for some apparent differences in their outputs.

Author

Laurent Gatto

Examples

f <- msdata::ident(full.names = TRUE, pattern = "TMT")
basename(f)

#> [1] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzid"

## mzR parser (default)
readPSMs(f)

#> DataFrame with 5802 rows and 35 columns
#>                    sequence             spectrumID chargeState      rank
#>                 <character>            <character>   <integer> <integer>
#> 1              RQCRTDFLNYLR controllerType=0 con..           3         1
#> 2    ESVALADQVTCVDWRNRKATKK controllerType=0 con..           2         1
#> 3              KELLCLAMQIIR controllerType=0 con..           2         1
#> 4    QRMARTSDKQQSIRFLERLCGR controllerType=0 con..           3         1
#> 5    KDEGSTEPLKVVRDMTDAIC.. controllerType=0 con..           3         1
#> ...                     ...                    ...         ...       ...
#> 5798              LLLARLFLK controllerType=0 con..           2         1
#> 5799                LFSLLKK controllerType=0 con..           2         1
#> 5800              LAATKKITK controllerType=0 con..           2         1
#> 5801               RLIPVVKR controllerType=0 con..           2         1
#> 5802                 IVKFRR controllerType=0 con..           2         1
#>      passThreshold experimentalMassToCharge calculatedMassToCharge  peptideRef
#>          <logical>                <numeric>              <numeric> <character>
#> 1             TRUE                  548.286                547.947       Pep77
#> 2             TRUE                 1288.153               1288.174      Pep108
#> 3             TRUE                  744.411                744.425      Pep136
#> 4             TRUE                  913.164                912.813      Pep163
#> 5             TRUE                  926.810                926.471      Pep169
#> ...            ...                      ...                    ...         ...
#> 5798          TRUE                  543.874                543.874     Pep4927
#> 5799          TRUE                  425.290                424.784     Pep4842
#> 5800          TRUE                  487.834                487.324     Pep4891
#> 5801          TRUE                  491.343                490.840     Pep4885
#> 5802          TRUE                  410.281                409.772     Pep4861
#>         modNum   isDecoy        post         pre     start       end
#>      <integer> <logical> <character> <character> <integer> <integer>
#> 1            1     FALSE           V           R       574       585
#> 2            1     FALSE           G           R        69        90
#> 3            1      TRUE           Q           R       131       142
#> 4            1     FALSE           D           R       182       203
#> 5            1      TRUE           R           R       135       158
#> ...        ...       ...         ...         ...       ...       ...
#> 5798         0     FALSE           P           R       449       457
#> 5799         0     FALSE           R           R        40        46
#> 5800         0     FALSE           -           R       312       320
#> 5801         0      TRUE           G           K       160       167
#> 5802         0     FALSE           R           K        74        79
#>      DatabaseAccess DBseqLength DatabaseSeq    DatabaseDescription
#>         <character>   <integer> <character>            <character>
#> 1           ECA2006        1295             ECA2006 ATP-dependen..
#> 2           ECA1676         110             ECA1676 putative gro..
#> 3       XXX_ECA2855         157                                   
#> 4           ECA3009         437             ECA3009 putative cop..
#> 5       XXX_ECA0048         501                                   
#> ...             ...         ...         ...                    ...
#> 5798        ECA2534         636             ECA2534 ABC transpor..
#> 5799        ECA1583          83             ECA1583 putative int..
#> 5800        ECA0610         320             ECA0610 LysR-family ..
#> 5801    XXX_ECA3377         211                                   
#> 5802        ECA0690         103             ECA0690 50S ribosoma..
#>      scan.number.s. acquisitionNum           spectrumFile
#>           <numeric>      <numeric>            <character>
#> 1              2949           2949 TMT_Erwinia_1uLSike_..
#> 2              6534           6534 TMT_Erwinia_1uLSike_..
#> 3              5674           5674 TMT_Erwinia_1uLSike_..
#> 4              4782           4782 TMT_Erwinia_1uLSike_..
#> 5              5839           5839 TMT_Erwinia_1uLSike_..
#> ...             ...            ...                    ...
#> 5798           2533           2533 TMT_Erwinia_1uLSike_..
#> 5799           2645           2645 TMT_Erwinia_1uLSike_..
#> 5800           1699           1699 TMT_Erwinia_1uLSike_..
#> 5801           6739           6739 TMT_Erwinia_1uLSike_..
#> 5802           1634           1634 TMT_Erwinia_1uLSike_..
#>                      idFile MS.GF.RawScore MS.GF.DeNovoScore MS.GF.SpecEValue
#>                 <character>      <numeric>         <numeric>        <numeric>
#> 1    TMT_Erwinia_1uLSike_..             10               101      4.61712e-08
#> 2    TMT_Erwinia_1uLSike_..             12               121      7.25588e-08
#> 3    TMT_Erwinia_1uLSike_..              8                74      9.34102e-08
#> 4    TMT_Erwinia_1uLSike_..             -5               160      1.27073e-07
#> 5    TMT_Erwinia_1uLSike_..              8               241      1.31824e-07
#> ...                     ...            ...               ...              ...
#> 5798 TMT_Erwinia_1uLSike_..            -90               118      0.004498388
#> 5799 TMT_Erwinia_1uLSike_..           -102                43      0.006621415
#> 5800 TMT_Erwinia_1uLSike_..            -66                90      0.001297466
#> 5801 TMT_Erwinia_1uLSike_..            -91                16      0.001965771
#> 5802 TMT_Erwinia_1uLSike_..            -44                54      0.000919113
#>      MS.GF.EValue MS.GF.QValue MS.GF.PepQValue modPeptideRef         modName
#>         <numeric>    <numeric>       <numeric>   <character>     <character>
#> 1        0.132198     0.525424        0.549020         Pep77 Carbamidomethyl
#> 2        0.208748     0.610390        0.623188        Pep108 Carbamidomethyl
#> 3        0.267453     0.625000        0.636364        Pep136 Carbamidomethyl
#> 4        0.365583     0.716981        0.724490        Pep163 Carbamidomethyl
#> 5        0.379461     0.736364        0.745098        Pep169 Carbamidomethyl
#> ...           ...          ...             ...           ...             ...
#> 5798     12838.72            1               1            NA              NA
#> 5799     18816.18            1               1            NA              NA
#> 5800      3703.06            1               1            NA              NA
#> 5801      5600.99            1               1            NA              NA
#> 5802      2587.26            1               1            NA              NA
#>        modMass modLocation subOriginalResidue subReplacementResidue subLocation
#>      <numeric>   <integer>        <character>           <character>   <integer>
#> 1      57.0215           3                 NA                    NA          NA
#> 2      57.0215          11                 NA                    NA          NA
#> 3      57.0215           5                 NA                    NA          NA
#> 4      57.0215          20                 NA                    NA          NA
#> 5      57.0215          20                 NA                    NA          NA
#> ...        ...         ...                ...                   ...         ...
#> 5798        NA          NA                 NA                    NA          NA
#> 5799        NA          NA                 NA                    NA          NA
#> 5800        NA          NA                 NA                    NA          NA
#> 5801        NA          NA                 NA                    NA          NA
#> 5802        NA          NA                 NA                    NA          NA

## mzID parser
readPSMs(f, parser = "mzID")

#> Loading required namespace: mzID

#> reading TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzid... DONE!

#> DataFrame with 5759 rows and 30 columns
#>                  spectrumid scan number(s) acquisitionnum passthreshold
#>                 <character>      <numeric>      <numeric>     <logical>
#> 1    controllerType=0 con..           5782           5782          TRUE
#> 2    controllerType=0 con..           6037           6037          TRUE
#> 3    controllerType=0 con..           5235           5235          TRUE
#> 4    controllerType=0 con..           5397           5397          TRUE
#> 5    controllerType=0 con..           6075           6075          TRUE
#> ...                     ...            ...            ...           ...
#> 5339 controllerType=0 con..           2758           2758          TRUE
#> 5340 controllerType=0 con..           3477           3477          TRUE
#> 5341 controllerType=0 con..           2913           2913          TRUE
#> 5342 controllerType=0 con..           3275           3275          TRUE
#> 5343 controllerType=0 con..           3017           3017          TRUE
#>           rank calculatedmasstocharge experimentalmasstocharge chargestate
#>      <numeric>              <numeric>                <numeric>   <numeric>
#> 1            1               1080.232                 1080.233           3
#> 2            1               1002.212                 1002.209           3
#> 3            1               1189.280                 1189.284           3
#> 4            1                960.536                  960.536           3
#> 5            1               1264.342                 1264.341           3
#> ...        ...                    ...                      ...         ...
#> 5339         1                592.377                  592.717           3
#> 5340         1                554.864                  554.867           2
#> 5341         1                483.821                  483.829           2
#> 5342         1                552.369                  552.880           2
#> 5343         1                483.821                  483.829           2
#>      ms-gf:denovoscore ms-gf:evalue ms-gf:pepqvalue ms-gf:qvalue ms-gf:rawscore
#>              <numeric>    <numeric>       <numeric>    <numeric>      <numeric>
#> 1                  174  1.08603e-20               0            0            147
#> 2                  245  1.98877e-19               0            0            214
#> 3                  264  5.12965e-19               0            0            211
#> 4                  178  5.16357e-18               0            0            154
#> 5                  252  4.35691e-17               0            0            188
#> ...                ...          ...             ...          ...            ...
#> 5339               205      20105.6               1            1            -72
#> 5340               130      20987.5               1            1           -102
#> 5341                52      24234.7               1            1           -141
#> 5342               157      29543.9               1            1           -111
#> 5343                64      31157.9               1            1           -133
#>      ms-gf:specevalue assumeddissociationmethod isotopeerror   isdecoy
#>             <numeric>               <character>  <character> <logical>
#> 1         3.76483e-27                       HCD            0     FALSE
#> 2         6.90263e-26                       HCD            0     FALSE
#> 3         1.77879e-25                       HCD            0     FALSE
#> 4         1.79254e-24                       HCD            0     FALSE
#> 5         1.51036e-23                       HCD            0     FALSE
#> ...               ...                       ...          ...       ...
#> 5339       0.00700092                       HCD            1     FALSE
#> 5340       0.00733647                       HCD            0      TRUE
#> 5341       0.00848043                       HCD            0      TRUE
#> 5342       0.01033827                       HCD            1      TRUE
#> 5343       0.01090306                       HCD            0      TRUE
#>             post         pre       end     start   accession    length
#>      <character> <character> <numeric> <numeric> <character> <numeric>
#> 1              S           R        84        50     ECA1932       155
#> 2              R           K       315       288     ECA1147       434
#> 3              A           R       224       192     ECA0013       295
#> 4              -           R       290       264     ECA1731       290
#> 5              F           R       153       119     ECA1443       298
#> ...          ...         ...       ...       ...         ...       ...
#> 5339           E           -        17         1     ECA3702        85
#> 5340           A           R       387       377 XXX_ECA1426       598
#> 5341           I           R        65        56 XXX_ECA0367       138
#> 5342           K           K        51        42 XXX_ECA3548       574
#> 5343           I           R        65        56 XXX_ECA0367       138
#>                 description                 pepseq  modified modification
#>                 <character>            <character> <logical>  <character>
#> 1    outer membrane lipop.. PVQIQAGEDSNVIGALGGAV..     FALSE           NA
#> 2            trigger factor TQVLDGLINANDIEVPVALI..     FALSE           NA
#> 3    ribose-binding perip.. TKGLNVMQNLLTAHPDVQAV..     FALSE           NA
#> 4                 flagellin SQILQQAGTSVLSQANQVPQ..     FALSE           NA
#> 5    UTP--glucose-1-phosp.. PIIGDNPFVVVLPDVVLDES..     FALSE           NA
#> ...                     ...                    ...       ...          ...
#> 5339   hypothetical protein      VAKIIISVQPLPAAQVK     FALSE           NA
#> 5340                     NA            VGVGALIVPRK     FALSE           NA
#> 5341                     NA             LLAIGAPLAK     FALSE           NA
#> 5342                     NA             ILHVVIPKGK     FALSE           NA
#> 5343                     NA             LLAIGAPLAK     FALSE           NA
#>                      idFile           spectrumFile           databaseFile
#>                 <character>            <character>            <character>
#> 1    TMT_Erwinia_1uLSike_.. TMT_Erwinia_1uLSike_.. erwinia_carotovora.f..
#> 2    TMT_Erwinia_1uLSike_.. TMT_Erwinia_1uLSike_.. erwinia_carotovora.f..
#> 3    TMT_Erwinia_1uLSike_.. TMT_Erwinia_1uLSike_.. erwinia_carotovora.f..
#> 4    TMT_Erwinia_1uLSike_.. TMT_Erwinia_1uLSike_.. erwinia_carotovora.f..
#> 5    TMT_Erwinia_1uLSike_.. TMT_Erwinia_1uLSike_.. erwinia_carotovora.f..
#> ...                     ...                    ...                    ...
#> 5339 TMT_Erwinia_1uLSike_.. TMT_Erwinia_1uLSike_.. erwinia_carotovora.f..
#> 5340 TMT_Erwinia_1uLSike_.. TMT_Erwinia_1uLSike_.. erwinia_carotovora.f..
#> 5341 TMT_Erwinia_1uLSike_.. TMT_Erwinia_1uLSike_.. erwinia_carotovora.f..
#> 5342 TMT_Erwinia_1uLSike_.. TMT_Erwinia_1uLSike_.. erwinia_carotovora.f..
#> 5343 TMT_Erwinia_1uLSike_.. TMT_Erwinia_1uLSike_.. erwinia_carotovora.f..