Filter out unreliable PSMs.

A function to filter out PSMs matching to the decoy database, of rank greater than one and matching non-proteotypic peptides.

filterPSMs(
  x,
  decoy = "isDecoy",
  rank = "rank",
  accession = "DatabaseAccess",
  spectrumID = "spectrumID",
  verbose = TRUE
)

Arguments

x
A DataFrame containing PSMs.

decoy
The column name defining whether entries match the decoy database. Default is "isDecoy". The column should be a logical and only PSMs holding a FALSE are retained. Ignored is set to NULL.

rank
The column name holding the rank of the PSM. Default is "rank". This column should be a numeric and only PSMs having rank equal to 1 are retained. Ignored is set to NULL.

accession
The column name holding the protein (groups) accession. Default is "DatabaseAccess". Ignored is set to NULL.

spectrumID
The name of the spectrum identifier column. Default is spectrumID.

verbose
logical(1) setting the verbosity flag.

Value

A new DataFrame with filtered out peptides and with the same columns as the input x.

Details

The PSMs should be stored in a DataFrame (or data.frame or tibble) such as those produced by readPSMs(). Note that this function should be used before reducing the PSM table.

Author

Laurent Gatto

Examples

f <- msdata::ident(full.names = TRUE, pattern = "TMT")
basename(f)

#> [1] "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01-20141210.mzid"

id <- readPSMs(f)

#> Loading required namespace: mzR

filterPSMs(id)

#> Starting with 5802 PSMs:

#>  removed 2896 decoy hits

#>  removed 155 PSMs with rank > 1

#>  removed 85 non-proteotypic peptides

#> 2666 PSMs left.

#> DataFrame with 2666 rows and 35 columns
#>                    sequence             spectrumID chargeState      rank
#>                 <character>            <character>   <integer> <integer>
#> 1              RQCRTDFLNYLR controllerType=0 con..           3         1
#> 2    ESVALADQVTCVDWRNRKATKK controllerType=0 con..           2         1
#> 3    QRMARTSDKQQSIRFLERLCGR controllerType=0 con..           3         1
#> 4    DGGPAIYGHERVGRNAKKFK.. controllerType=0 con..           3         1
#> 5    QRMARTSDKQQSIRFLERLCGR controllerType=0 con..           2         1
#> ...                     ...                    ...         ...       ...
#> 2662                KKRAKVK controllerType=0 con..           2         1
#> 2663              LLLARLFLK controllerType=0 con..           2         1
#> 2664                LFSLLKK controllerType=0 con..           2         1
#> 2665              LAATKKITK controllerType=0 con..           2         1
#> 2666                 IVKFRR controllerType=0 con..           2         1
#>      passThreshold experimentalMassToCharge calculatedMassToCharge  peptideRef
#>          <logical>                <numeric>              <numeric> <character>
#> 1             TRUE                  548.286                547.947       Pep77
#> 2             TRUE                 1288.153               1288.174      Pep108
#> 3             TRUE                  913.164                912.813      Pep163
#> 4             TRUE                  969.194                968.856      Pep206
#> 5             TRUE                 1369.238               1368.717      Pep163
#> ...            ...                      ...                    ...         ...
#> 2662          TRUE                  429.800                429.306     Pep4913
#> 2663          TRUE                  543.874                543.874     Pep4927
#> 2664          TRUE                  425.290                424.784     Pep4842
#> 2665          TRUE                  487.834                487.324     Pep4891
#> 2666          TRUE                  410.281                409.772     Pep4861
#>         modNum   isDecoy        post         pre     start       end
#>      <integer> <logical> <character> <character> <integer> <integer>
#> 1            1     FALSE           V           R       574       585
#> 2            1     FALSE           G           R        69        90
#> 3            1     FALSE           D           R       182       203
#> 4            1     FALSE           S           K       310       334
#> 5            1     FALSE           D           R       182       203
#> ...        ...       ...         ...         ...       ...       ...
#> 2662         0     FALSE           N           R       547       553
#> 2663         0     FALSE           P           R       449       457
#> 2664         0     FALSE           R           R        40        46
#> 2665         0     FALSE           -           R       312       320
#> 2666         0     FALSE           R           K        74        79
#>      DatabaseAccess DBseqLength DatabaseSeq    DatabaseDescription
#>         <character>   <integer> <character>            <character>
#> 1           ECA2006        1295             ECA2006 ATP-dependen..
#> 2           ECA1676         110             ECA1676 putative gro..
#> 3           ECA3009         437             ECA3009 putative cop..
#> 4           ECA1420         477             ECA1420 undecaprenyl..
#> 5           ECA3009         437             ECA3009 putative cop..
#> ...             ...         ...         ...                    ...
#> 2662        ECA2881         610             ECA2881 excinuclease..
#> 2663        ECA2534         636             ECA2534 ABC transpor..
#> 2664        ECA1583          83             ECA1583 putative int..
#> 2665        ECA0610         320             ECA0610 LysR-family ..
#> 2666        ECA0690         103             ECA0690 50S ribosoma..
#>      scan.number.s. acquisitionNum           spectrumFile
#>           <numeric>      <numeric>            <character>
#> 1              2949           2949 TMT_Erwinia_1uLSike_..
#> 2              6534           6534 TMT_Erwinia_1uLSike_..
#> 3              4782           4782 TMT_Erwinia_1uLSike_..
#> 4              6118           6118 TMT_Erwinia_1uLSike_..
#> 5              4757           4757 TMT_Erwinia_1uLSike_..
#> ...             ...            ...                    ...
#> 2662           1079           1079 TMT_Erwinia_1uLSike_..
#> 2663           2533           2533 TMT_Erwinia_1uLSike_..
#> 2664           2645           2645 TMT_Erwinia_1uLSike_..
#> 2665           1699           1699 TMT_Erwinia_1uLSike_..
#> 2666           1634           1634 TMT_Erwinia_1uLSike_..
#>                      idFile MS.GF.RawScore MS.GF.DeNovoScore MS.GF.SpecEValue
#>                 <character>      <numeric>         <numeric>        <numeric>
#> 1    TMT_Erwinia_1uLSike_..             10               101      4.61712e-08
#> 2    TMT_Erwinia_1uLSike_..             12               121      7.25588e-08
#> 3    TMT_Erwinia_1uLSike_..             -5               160      1.27073e-07
#> 4    TMT_Erwinia_1uLSike_..              7               214      1.72203e-07
#> 5    TMT_Erwinia_1uLSike_..             21               196      1.76840e-07
#> ...                     ...            ...               ...              ...
#> 2662 TMT_Erwinia_1uLSike_..            -76                48      0.002470319
#> 2663 TMT_Erwinia_1uLSike_..            -90               118      0.004498388
#> 2664 TMT_Erwinia_1uLSike_..           -102                43      0.006621415
#> 2665 TMT_Erwinia_1uLSike_..            -66                90      0.001297466
#> 2666 TMT_Erwinia_1uLSike_..            -44                54      0.000919113
#>      MS.GF.EValue MS.GF.QValue MS.GF.PepQValue modPeptideRef         modName
#>         <numeric>    <numeric>       <numeric>   <character>     <character>
#> 1        0.132198     0.525424        0.549020         Pep77 Carbamidomethyl
#> 2        0.208748     0.610390        0.623188        Pep108 Carbamidomethyl
#> 3        0.365583     0.716981        0.724490        Pep163 Carbamidomethyl
#> 4        0.495817     0.837037        0.856000        Pep206 Carbamidomethyl
#> 5        0.508761     0.837037        0.724490        Pep163 Carbamidomethyl
#> ...           ...          ...             ...           ...             ...
#> 2662      7019.94            1               1            NA              NA
#> 2663     12838.72            1               1            NA              NA
#> 2664     18816.18            1               1            NA              NA
#> 2665      3703.06            1               1            NA              NA
#> 2666      2587.26            1               1            NA              NA
#>        modMass modLocation subOriginalResidue subReplacementResidue subLocation
#>      <numeric>   <integer>        <character>           <character>   <integer>
#> 1      57.0215           3                 NA                    NA          NA
#> 2      57.0215          11                 NA                    NA          NA
#> 3      57.0215          20                 NA                    NA          NA
#> 4      57.0215          21                 NA                    NA          NA
#> 5      57.0215          20                 NA                    NA          NA
#> ...        ...         ...                ...                   ...         ...
#> 2662        NA          NA                 NA                    NA          NA
#> 2663        NA          NA                 NA                    NA          NA
#> 2664        NA          NA                 NA                    NA          NA
#> 2665        NA          NA                 NA                    NA          NA
#> 2666        NA          NA                 NA                    NA          NA